# controllers/api/campaign.go
package api

import (
	"encoding/json"
	"fmt"
	"net/http"
	"strconv"

	ctx "github.com/gophish/gophish/context"
	log "github.com/gophish/gophish/logger"
	"github.com/gophish/gophish/models"
	"github.com/gorilla/mux"
	"github.com/jinzhu/gorm"
)

// Campaigns returns a list of campaigns if requested via GET.
// If requested via POST, APICampaigns creates a new campaign and returns a reference to it.
func (as *Server) Campaigns(w http.ResponseWriter, r *http.Request) {
	switch {
	case r.Method == "GET":
		cs, err := models.GetCampaigns(ctx.Get(r, "user_id").(int64))
		if err != nil {
			log.Error(err)
		}
		JSONResponse(w, cs, http.StatusOK)
	//POST: Create a new campaign and return it as JSON
	case r.Method == "POST":
		c := models.Campaign{}
		// Put the request into a campaign
		err := json.NewDecoder(r.Body).Decode(&c)
		if err != nil {
			JSONResponse(w, models.Response{Success: false, Message: "Invalid JSON structure"}, http.StatusBadRequest)
			return
		}
		err = models.PostCampaign(&c, ctx.Get(r, "user_id").(int64))
		if err != nil {
			JSONResponse(w, models.Response{Success: false, Message: err.Error()}, http.StatusBadRequest)
			return
		}
		// If the campaign is scheduled to launch immediately, send it to the worker.
		// Otherwise, the worker will pick it up at the scheduled time
		if c.Status == models.CampaignInProgress {
			go as.worker.LaunchCampaign(c)
		}
		JSONResponse(w, c, http.StatusCreated)
	}
}

// CampaignsSummary returns the summary for the current user's campaigns
func (as *Server) CampaignsSummary(w http.ResponseWriter, r *http.Request) {
	switch {
	case r.Method == "GET":
		cs, err := models.GetCampaignSummaries(ctx.Get(r, "user_id").(int64))
		if err != nil {
			log.Error(err)
			JSONResponse(w, models.Response{Success: false, Message: err.Error()}, http.StatusInternalServerError)
			return
		}
		JSONResponse(w, cs, http.StatusOK)
	}
}

// Campaign returns details about the requested campaign. If the campaign is not
// valid, APICampaign returns null.
func (as *Server) Campaign(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	id, _ := strconv.ParseInt(vars["id"], 0, 64)
	c, err := models.GetCampaign(id, ctx.Get(r, "user_id").(int64))
	if err != nil {
		log.Error(err)
		JSONResponse(w, models.Response{Success: false, Message: "Campaign not found"}, http.StatusNotFound)
		return
	}
	switch {
	case r.Method == "GET":
		JSONResponse(w, c, http.StatusOK)
	case r.Method == "DELETE":
		err = models.DeleteCampaign(id)
		if err != nil {
			JSONResponse(w, models.Response{Success: false, Message: "Error deleting campaign"}, http.StatusInternalServerError)
			return
		}
		JSONResponse(w, models.Response{Success: true, Message: "Campaign deleted successfully!"}, http.StatusOK)
	}
}

// CampaignResults returns just the results for a given campaign to
// significantly reduce the information returned.
func (as *Server) CampaignResults(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	id, _ := strconv.ParseInt(vars["id"], 0, 64)
	cr, err := models.GetCampaignResults(id, ctx.Get(r, "user_id").(int64))
	if err != nil {
		log.Error(err)
		JSONResponse(w, models.Response{Success: false, Message: "Campaign not found"}, http.StatusNotFound)
		return
	}
	if r.Method == "GET" {
		JSONResponse(w, cr, http.StatusOK)
		return
	}
}

// CampaignSummary returns the summary for a given campaign.
func (as *Server) CampaignSummary(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	id, _ := strconv.ParseInt(vars["id"], 0, 64)
	switch {
	case r.Method == "GET":
		cs, err := models.GetCampaignSummary(id, ctx.Get(r, "user_id").(int64))
		if err != nil {
			if err == gorm.ErrRecordNotFound {
				JSONResponse(w, models.Response{Success: false, Message: "Campaign not found"}, http.StatusNotFound)
			} else {
				JSONResponse(w, models.Response{Success: false, Message: err.Error()}, http.StatusInternalServerError)
			}
			log.Error(err)
			return
		}
		JSONResponse(w, cs, http.StatusOK)
	}
}

// CampaignComplete effectively "ends" a campaign.
// Future phishing emails clicked will return a simple "404" page.
func (as *Server) CampaignComplete(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	id, _ := strconv.ParseInt(vars["id"], 0, 64)
	switch {
	case r.Method == "GET":
		err := models.CompleteCampaign(id, ctx.Get(r, "user_id").(int64))
		if err != nil {
			JSONResponse(w, models.Response{Success: false, Message: "Error completing campaign"}, http.StatusInternalServerError)
			return
		}
		JSONResponse(w, models.Response{Success: true, Message: "Campaign completed successfully!"}, http.StatusOK)
	}
}

// CampaignReopen reverts a completed campaign to "In progress" so events are
// processed again. Emails that were dropped from the send queue when the
// campaign was completed are not re-sent.
func (as *Server) CampaignReopen(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	id, _ := strconv.ParseInt(vars["id"], 0, 64)
	switch {
	case r.Method == "GET":
		err := models.ReopenCampaign(id, ctx.Get(r, "user_id").(int64))
		if err != nil {
			JSONResponse(w, models.Response{Success: false, Message: "Error reopening campaign"}, http.StatusInternalServerError)
			return
		}
		JSONResponse(w, models.Response{Success: true, Message: "Campaign reopened successfully!"}, http.StatusOK)
	}
}

// CampaignSpamReport fires the n8n spam-report webhook for a campaign. The
// webhook is sent in the background (fire-and-forget): this handler returns as
// soon as the request is queued without waiting for the webhook's response.
// It accepts POST /campaigns/{id}/spam-report.
func (as *Server) CampaignSpamReport(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodPost {
		JSONResponse(w, models.Response{Success: false, Message: "Method not allowed"}, http.StatusMethodNotAllowed)
		return
	}
	vars := mux.Vars(r)
	id, _ := strconv.ParseInt(vars["id"], 0, 64)
	c, err := models.GetCampaign(id, ctx.Get(r, "user_id").(int64))
	if err != nil {
		JSONResponse(w, models.Response{Success: false, Message: "Campaign not found"}, http.StatusNotFound)
		return
	}
	if err := models.SendSpamReportWebhook(&c); err != nil {
		JSONResponse(w, models.Response{Success: false, Message: err.Error()}, http.StatusInternalServerError)
		return
	}
	JSONResponse(w, models.Response{Success: true, Message: "Spam report webhook sent"}, http.StatusOK)
}

// ResultReported marks a single recipient in a campaign as having reported the
// phishing email, setting the Reported flag and adding an "Email Reported" event
// to the campaign timeline. This is the authenticated, admin-facing counterpart
// to the public phishing-server report handler: it lets the SOC manually flag a
// recipient who reported out-of-band (e.g. by phone or forwarded email) rather
// than through the automated IMAP/report mechanism.
// It accepts POST /campaigns/{id}/results/{rid}/report.
func (as *Server) ResultReported(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodPost {
		JSONResponse(w, models.Response{Success: false, Message: "Method not allowed"}, http.StatusMethodNotAllowed)
		return
	}
	vars := mux.Vars(r)
	id, _ := strconv.ParseInt(vars["id"], 0, 64)
	rid := vars["rid"]
	c, err := models.GetCampaign(id, ctx.Get(r, "user_id").(int64))
	if err != nil {
		JSONResponse(w, models.Response{Success: false, Message: "Campaign not found"}, http.StatusNotFound)
		return
	}
	result, err := models.GetResult(rid)
	if err != nil {
		JSONResponse(w, models.Response{Success: false, Message: fmt.Sprintf("result %q not found", rid)}, http.StatusNotFound)
		return
	}
	if result.CampaignId != c.Id {
		JSONResponse(w, models.Response{Success: false, Message: fmt.Sprintf("result %q does not belong to this campaign", rid)}, http.StatusBadRequest)
		return
	}
	// Don't create a duplicate timeline event if the recipient is already
	// flagged as reported.
	if result.Reported {
		JSONResponse(w, models.Response{Success: true, Message: "Recipient already marked as reported"}, http.StatusOK)
		return
	}
	if err := result.HandleEmailReport(models.EventDetails{}); err != nil {
		log.Error(err)
		JSONResponse(w, models.Response{Success: false, Message: "Error marking recipient as reported"}, http.StatusInternalServerError)
		return
	}
	JSONResponse(w, models.Response{Success: true, Message: "Recipient marked as reported"}, http.StatusOK)
}

// buildResendMailLog validates that the result identified by rid belongs to the
// given campaign and prepares a fresh, locked maillog ready to be queued for an
// immediate resend.
func buildResendMailLog(c *models.Campaign, rid string) (*models.MailLog, error) {
	result, err := models.GetResult(rid)
	if err != nil {
		return nil, fmt.Errorf("result %q not found", rid)
	}
	if result.CampaignId != c.Id {
		return nil, fmt.Errorf("result %q does not belong to this campaign", rid)
	}
	return models.GenerateResendMailLog(c, &result)
}

// queueResends caches the campaign mail context onto the maillogs (so the
// mailer doesn't re-fetch it per email) and hands them to the worker to be
// sent immediately.
func (as *Server) queueResends(c *models.Campaign, mailLogs []*models.MailLog) {
	if mailCtx, err := models.GetCampaignMailContext(c.Id, c.UserId); err == nil {
		for _, ml := range mailLogs {
			ml.CacheCampaign(&mailCtx)
		}
	}
	go as.worker.SendEmails(mailLogs)
}

// ResendResult re-sends the phishing email to a single recipient in a campaign.
// It accepts POST /campaigns/{id}/results/{rid}/resend.
func (as *Server) ResendResult(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodPost {
		JSONResponse(w, models.Response{Success: false, Message: "Method not allowed"}, http.StatusMethodNotAllowed)
		return
	}
	vars := mux.Vars(r)
	id, _ := strconv.ParseInt(vars["id"], 0, 64)
	rid := vars["rid"]
	c, err := models.GetCampaign(id, ctx.Get(r, "user_id").(int64))
	if err != nil {
		JSONResponse(w, models.Response{Success: false, Message: "Campaign not found"}, http.StatusNotFound)
		return
	}
	ml, err := buildResendMailLog(&c, rid)
	if err != nil {
		JSONResponse(w, models.Response{Success: false, Message: err.Error()}, http.StatusBadRequest)
		return
	}
	as.queueResends(&c, []*models.MailLog{ml})
	JSONResponse(w, models.Response{Success: true, Message: "Email queued to be resent"}, http.StatusOK)
}

// ResendResults re-sends the phishing email to multiple recipients in a
// campaign. It accepts POST /campaigns/{id}/results/resend with a JSON body of
// the form {"ids": ["abc1234", ...]}.
func (as *Server) ResendResults(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodPost {
		JSONResponse(w, models.Response{Success: false, Message: "Method not allowed"}, http.StatusMethodNotAllowed)
		return
	}
	vars := mux.Vars(r)
	id, _ := strconv.ParseInt(vars["id"], 0, 64)
	c, err := models.GetCampaign(id, ctx.Get(r, "user_id").(int64))
	if err != nil {
		JSONResponse(w, models.Response{Success: false, Message: "Campaign not found"}, http.StatusNotFound)
		return
	}
	payload := struct {
		IDs []string `json:"ids"`
	}{}
	err = json.NewDecoder(r.Body).Decode(&payload)
	if err != nil {
		JSONResponse(w, models.Response{Success: false, Message: "Invalid JSON structure"}, http.StatusBadRequest)
		return
	}
	if len(payload.IDs) == 0 {
		JSONResponse(w, models.Response{Success: false, Message: "No recipients selected"}, http.StatusBadRequest)
		return
	}
	mailLogs := []*models.MailLog{}
	for _, rid := range payload.IDs {
		ml, err := buildResendMailLog(&c, rid)
		if err != nil {
			log.Error(err)
			continue
		}
		mailLogs = append(mailLogs, ml)
	}
	if len(mailLogs) == 0 {
		JSONResponse(w, models.Response{Success: false, Message: "No valid recipients to resend"}, http.StatusBadRequest)
		return
	}
	as.queueResends(&c, mailLogs)
	JSONResponse(w, models.Response{Success: true, Message: fmt.Sprintf("Queued %d email(s) to be resent", len(mailLogs))}, http.StatusOK)
}