# static/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gophish API (Yazamco SOC Edition)",
    "description": "REST API for this Gophish instance, including all Yazamco-specific extensions: **Companies** (Hudu-backed organizations), **Domains** (n8n / Cloudflare / Microsoft 365 provisioning), **template rotation pools**, **per-recipient email subjects**, **resend** endpoints and the **spam-report** webhook trigger.\n\n## Authentication\nEvery endpoint requires your personal API key, found under **Account Settings**. Supply it one of two ways:\n\n- `Authorization: Bearer <API_KEY>` header (recommended)\n- `?api_key=<API_KEY>` query parameter\n\n## Permissions\n- Users with the **user** role can manage only the objects they own; **admin** users see everything.\n- Endpoints marked *admin only* require the `modify_system` permission (User Management, Webhooks).\n- Accounts in *view-only* mode may only issue GET requests.\n\n## Conventions\n- All write endpoints accept and return `application/json` (except group CSV import, which is `multipart/form-data`).\n- Dates are RFC 3339 / ISO-8601 with timezone, e.g. `2026-06-04T09:30:00+03:00`. The zero date `0001-01-01T00:00:00Z` means \"not set\".\n- Errors return a `GenericResponse` envelope: `{\"message\": \"...\", \"success\": false, \"data\": null}`.",
    "version": "1.1-yazamco",
    "contact": {
      "name": "Yazamco Pro Cyber"
    },
    "license": {
      "name": "MIT",
...
    }
  },
  "servers": [
    {
      "url": "/api",
...
    }
  ],
  "security": [
    { "BearerAuth": [] },
    { "ApiKeyQuery": [] }
  ],
...
  "paths": {
    "/campaigns/": {
      "get": {
        "tags": ["Campaigns"],
        "summary": "List campaigns",
        "description": "Returns all campaigns visible to the authenticated user (admins see every user's campaigns), including full results and timeline for each. For dashboards prefer `/campaigns/summary`, which is much lighter.",
        "operationId": "getCampaigns",
        "responses": {
          "200": {
            "description": "A list of campaigns",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Campaign" } }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Campaigns"],
        "summary": "Create and launch a campaign",
        "description": "Creates a new campaign. Referenced objects (template(s), page, groups, sending profile) are looked up **by name** and must already exist.\n\n- If `launch_date` is omitted or in the past, the campaign launches **immediately**.\n- If `launch_date` is in the future the campaign is queued (`status: Queued`) and the background worker launches it on time.\n- If `send_by_date` is set, email sending is spread evenly between `launch_date` and `send_by_date`.\n- **Template rotation (Yazamco extension):** provide multiple templates in `templates[]` and each recipient is randomly assigned one of them at launch. `template` may be omitted when `templates` is provided; the first pool entry is kept in `template` for compatibility.\n- `company_id` attributes the campaign to a Company (used in reports and the spam-report webhook).",
        "operationId": "createCampaign",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CampaignRequest" },
              "examples": {
                "immediate": {
                  "summary": "Launch immediately, single template",
                  "value": {
                    "name": "Q3 Awareness - Acme",
                    "template": { "name": "Password Reset" },
                    "page": { "name": "O365 Login" },
                    "groups": [ { "name": "Acme Employees" } ],
                    "smtp": { "name": "Acme Mailer" },
                    "url": "https://login.acme-portal.com",
                    "company_id": 3
                  }
                },
                "scheduledRotation": {
                  "summary": "Scheduled launch with template rotation",
                  "value": {
                    "name": "Q3 Awareness - Rotation",
                    "templates": [ { "name": "Password Reset" }, { "name": "Shared Document" }, { "name": "Payroll Update" } ],
                    "page": { "name": "O365 Login" },
                    "groups": [ { "name": "Acme Employees" } ],
                    "smtp": { "name": "Acme Mailer" },
                    "url": "https://login.acme-portal.com",
                    "launch_date": "2026-06-10T08:00:00+03:00",
                    "send_by_date": "2026-06-10T12:00:00+03:00",
                    "company_id": 3
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Campaign created (and launched/queued)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Campaign" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/campaigns/{id}": {
      "parameters": [ { "$ref": "#/components/parameters/CampaignId" } ],
      "get": {
        "tags": ["Campaigns"],
        "summary": "Get a campaign",
        "description": "Returns full campaign details including results and the event timeline.",
        "operationId": "getCampaign",
        "responses": {
          "200": { "description": "The campaign", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Campaign" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Campaigns"],
        "summary": "Delete a campaign",
        "description": "Permanently deletes the campaign, its results, events and queued emails. **Irreversible.**",
        "operationId": "deleteCampaign",
        "responses": {
          "200": { "$ref": "#/components/responses/GenericSuccess" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/campaigns/{id}/results": {
      "parameters": [ { "$ref": "#/components/parameters/CampaignId" } ],
      "get": {
        "tags": ["Results"],
        "summary": "Get campaign results only",
        "description": "Returns just the per-recipient results and event timeline — significantly less data than fetching the full campaign (no template/page/SMTP bodies).",
        "operationId": "getCampaignResults",
        "responses": {
          "200": { "description": "Results and timeline", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CampaignResults" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/campaigns/{id}/summary": {
      "parameters": [ { "$ref": "#/components/parameters/CampaignId" } ],
      "get": {
        "tags": ["Campaigns"],
        "summary": "Get a campaign summary",
        "description": "Aggregate stats for a single campaign (sent / opened / clicked / submitted / reported / errors).",
        "operationId": "getCampaignSummary",
        "responses": {
          "200": { "description": "The summary", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CampaignSummary" } } } },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/campaigns/{id}/complete": {
      "parameters": [ { "$ref": "#/components/parameters/CampaignId" } ],
      "get": {
        "tags": ["Campaigns"],
        "summary": "Complete a campaign",
        "description": "Marks the campaign as **Completed**. No further emails are sent; subsequent clicks on phishing links return a 404 page. (Note: this is a GET request for historical compatibility.)",
        "operationId": "completeCampaign",
        "responses": {
          "200": { "$ref": "#/components/responses/GenericSuccess" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/campaigns/{id}/reopen": {
      "parameters": [ { "$ref": "#/components/parameters/CampaignId" } ],
      "get": {
        "tags": ["Campaigns"],
        "summary": "Reopen a completed campaign",
        "description": "Reverts a **Completed** campaign back to **In progress** so the phishing server resumes recording events (opens, clicks, submissions). Emails that were still queued when the campaign was completed are **not** re-sent. No-op if the campaign is not completed. (Note: this is a GET request for consistency with `/complete`.)",
        "operationId": "reopenCampaign",
        "responses": {
          "200": { "$ref": "#/components/responses/GenericSuccess" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/campaigns/{id}/results/resend": {
      "parameters": [ { "$ref": "#/components/parameters/CampaignId" } ],
      "post": {
        "tags": ["Results"],
        "summary": "Resend to multiple recipients (Yazamco extension)",
        "description": "Re-sends the phishing email to the selected recipients of a campaign. Each recipient keeps their originally-assigned template (rotation pools are respected). Invalid result IDs are skipped; the response reports how many emails were queued. Sending is asynchronous.",
        "operationId": "resendResults",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResendRequest" },
              "example": { "ids": ["b7c9d2e", "a1f4g8h"] }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Emails queued",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GenericResponse" }, "example": { "message": "Queued 2 email(s) to be resent", "success": true, "data": null } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      }
    },
    "/campaigns/{id}/results/{rid}/resend": {
      "parameters": [
        { "$ref": "#/components/parameters/CampaignId" },
        {
          "name": "rid",
          "in": "path",
          "required": true,
          "description": "The recipient's result ID (`Result.id`, e.g. `b7c9d2e`) — **not** a numeric row id.",
          "schema": { "type": "string" }
        }
      ],
      "post": {
        "tags": ["Results"],
        "summary": "Resend to a single recipient (Yazamco extension)",
        "description": "Re-sends the phishing email to one recipient. The result must belong to the campaign. Sending is asynchronous.",
        "operationId": "resendResult",
        "responses": {
          "200": {
            "description": "Email queued",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GenericResponse" }, "example": { "message": "Email queued to be resent", "success": true, "data": null } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" }
        }
      }
    },
...
    "/campaigns/{id}/spam-report": {
      "parameters": [ { "$ref": "#/components/parameters/CampaignId" } ],
      "post": {
        "tags": ["Results"],
        "summary": "Trigger the spam-report webhook (Yazamco extension)",
        "description": "Fires the n8n **spam-report** webhook for this campaign with the full per-recipient result set (including each recipient's rendered `email_subject`, assigned `template_id`, status and reported flag) plus the attributed Company (English + Hebrew names, Hudu customer id).\n\nThe webhook is **fire-and-forget**: a `200` here means the request was queued, not that n8n succeeded. Requires `spam_report_secret` in `config.json`; the payload is sent with the `x-techform-secret` header. Default endpoint: `https://n8n.yazamco.pro/webhook/gophish-check-spam-report-365` (override with `spam_report_url`).",
        "operationId": "campaignSpamReport",
        "responses": {
          "200": {
            "description": "Webhook queued",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GenericResponse" }, "example": { "message": "Spam report webhook sent", "success": true, "data": null } } }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" },
          "500": {
            "description": "Webhook not configured (`spam_report_secret` missing) or other failure",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GenericResponse" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Your personal API key from Account Settings, sent as `Authorization: Bearer <API_KEY>`."
      },
      "ApiKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "api_key",
        "description": "Your personal API key as a query parameter, e.g. `?api_key=<API_KEY>`."
      }
    },
    "parameters": {
...
      "CampaignId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "The numeric campaign ID",
        "schema": { "type": "integer", "format": "int64", "minimum": 1 }
      }
    },
    "responses": {
      "GenericSuccess": {
        "description": "Operation succeeded",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GenericResponse" } } }
      },
      "BadRequest": {
        "description": "Validation failed or malformed JSON — see `message`",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GenericResponse" } } }
      },
      "Unauthorized": {
        "description": "Missing or invalid API key",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GenericResponse" }, "example": { "message": "API Key not set", "success": false, "data": null } } }
      },
...
      "NotFound": {
        "description": "Object not found (or not owned by you)",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GenericResponse" } } }
      },
...
      "MethodNotAllowed": {
        "description": "Wrong HTTP method for this endpoint",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GenericResponse" } } }
      },
      "ServerError": {
        "description": "Internal server error",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GenericResponse" } } }
      }
    },
    "schemas": {
      "GenericResponse": {
        "type": "object",
        "description": "Standard response envelope for status messages and errors.",
        "properties": {
          "message": { "type": "string", "example": "Campaign deleted successfully!" },
          "success": { "type": "boolean" },
          "data": { "nullable": true, "description": "Optional payload (e.g. the new API key on `/reset`)" }
        }
      },
      "Campaign": {
        "type": "object",
        "description": "A phishing campaign with its full configuration, per-recipient results and event timeline.",
        "properties": {
          "id": { "type": "integer", "format": "int64", "readOnly": true },
          "name": { "type": "string", "example": "Q3 Awareness - Acme" },
          "created_date": { "type": "string", "format": "date-time", "readOnly": true },
          "launch_date": { "type": "string", "format": "date-time", "description": "When sending starts. Omit / past value = launch immediately." },
          "send_by_date": { "type": "string", "format": "date-time", "description": "Optional deadline; sending is spread evenly between launch_date and this time. Zero date = send as fast as possible." },
          "completed_date": { "type": "string", "format": "date-time", "readOnly": true },
          "template": { "$ref": "#/components/schemas/Template" },
          "templates": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Template" },
            "description": "**Yazamco extension** — optional template rotation pool. When more than one template is given, each recipient is randomly assigned one at launch; `template` holds the first pool entry for compatibility."
          },
          "page": { "$ref": "#/components/schemas/Page" },
          "status": { "$ref": "#/components/schemas/CampaignStatus" },
          "results": { "type": "array", "items": { "$ref": "#/components/schemas/Result" }, "readOnly": true },
          "groups": { "type": "array", "items": { "$ref": "#/components/schemas/Group" } },
          "timeline": { "type": "array", "items": { "$ref": "#/components/schemas/Event" }, "readOnly": true },
          "smtp": { "$ref": "#/components/schemas/SMTP" },
          "url": { "type": "string", "description": "Base URL of the phishing server; `{{.URL}}` in templates resolves to this plus `?rid=<result id>`", "example": "https://login.acme-portal.com" },
          "company_id": { "type": "integer", "format": "int64", "description": "ID of the Company this campaign is attributed to (0 = unassigned)" }
        }
      },
      "CampaignRequest": {
        "type": "object",
        "description": "Payload for creating a campaign. Referenced objects are matched **by name** and must already exist.",
        "required": ["name", "page", "groups", "smtp"],
        "properties": {
          "name": { "type": "string" },
          "template": { "type": "object", "properties": { "name": { "type": "string" } }, "description": "Email template reference (by name). Optional when `templates` is provided." },
          "templates": {
            "type": "array",
            "items": { "type": "object", "properties": { "name": { "type": "string" } } },
            "description": "**Yazamco extension** — rotation pool of template references. Each recipient gets a random one."
          },
          "page": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string" } }, "description": "Landing page reference (by name)" },
          "groups": {
            "type": "array",
            "minItems": 1,
            "items": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string" } } },
            "description": "Target group references (by name)"
          },
          "smtp": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string" } }, "description": "Sending profile reference (by name)" },
          "url": { "type": "string", "description": "Phishing server base URL used in `{{.URL}}`" },
          "launch_date": { "type": "string", "format": "date-time", "description": "Omit to launch immediately" },
          "send_by_date": { "type": "string", "format": "date-time", "description": "Must be after launch_date when set" },
          "company_id": { "type": "integer", "format": "int64", "description": "Optional Company attribution" }
        }
      },
      "CampaignStatus": {
        "type": "string",
        "description": "Campaign lifecycle status.",
        "enum": ["Created", "Queued", "In progress", "Emails Sent", "Completed"]
      },
      "CampaignResults": {
        "type": "object",
        "description": "Slim view of a campaign: results + timeline only.",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "status": { "$ref": "#/components/schemas/CampaignStatus" },
          "results": { "type": "array", "items": { "$ref": "#/components/schemas/Result" } },
          "timeline": { "type": "array", "items": { "$ref": "#/components/schemas/Event" } }
        }
      },
...
      "CampaignSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "created_date": { "type": "string", "format": "date-time" },
          "launch_date": { "type": "string", "format": "date-time" },
          "send_by_date": { "type": "string", "format": "date-time" },
          "completed_date": { "type": "string", "format": "date-time" },
          "status": { "$ref": "#/components/schemas/CampaignStatus" },
          "name": { "type": "string" },
          "company_id": { "type": "integer", "format": "int64", "description": "Attributed Company id (0 = unassigned)" },
          "company": { "type": "string", "description": "Attributed Company name" },
          "stats": { "$ref": "#/components/schemas/CampaignStats" }
        }
      },
...
      "ResendRequest": {
        "type": "object",
        "required": ["ids"],
        "properties": {
          "ids": {
            "type": "array",
            "minItems": 1,
            "items": { "type": "string" },
            "description": "Result ids (`Result.id` strings) of the recipients to resend to"
          }
        }
      },
      "Group": {
        "type": "object",
        "description": "A group of phishing targets.",
        "properties": {
          "id": { "type": "integer", "format": "int64", "readOnly": true },
          "name": { "type": "string", "example": "Acme Employees" },
          "modified_date": { "type": "string", "format": "date-time", "readOnly": true },
          "targets": { "type": "array", "items": { "$ref": "#/components/schemas/Target" } }
        }
      },
...
      "Template": {
        "type": "object",
        "description": "An email template. Subject, text and HTML support Go template variables (see `POST /templates/`).",
        "properties": {
          "id": { "type": "integer", "format": "int64", "readOnly": true },
          "name": { "type": "string", "example": "Password Reset" },
          "envelope_sender": { "type": "string", "description": "Optional SMTP MAIL FROM override; defaults to the sending profile's from_address", "example": "it-support@acme-portal.com" },
          "subject": { "type": "string", "example": "Action required: password expires today, {{.FirstName}}" },
          "text": { "type": "string", "description": "Plaintext body" },
          "html": { "type": "string", "description": "HTML body" },
          "modified_date": { "type": "string", "format": "date-time", "readOnly": true },
          "attachments": { "type": "array", "items": { "$ref": "#/components/schemas/Attachment" } }
        }
      },
...
      "Page": {
        "type": "object",
        "description": "A phishing landing page.",
        "properties": {
          "id": { "type": "integer", "format": "int64", "readOnly": true },
          "name": { "type": "string", "example": "O365 Login" },
          "html": { "type": "string", "description": "Page HTML. Forms are rewritten on save to post back to the phishing server." },
          "capture_credentials": { "type": "boolean", "description": "Record submitted form fields (passwords excluded unless capture_passwords)" },
          "capture_passwords": { "type": "boolean", "description": "Also record password fields — stored in **plaintext**" },
          "redirect_url": { "type": "string", "description": "Where the victim is redirected after submitting", "example": "https://office.com" },
          "modified_date": { "type": "string", "format": "date-time", "readOnly": true }
        }
      },
...
    }
  }
}